[refactor] move api calls to dedicate folder
